+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 Changelog
 =========
 
+
+0.2.3
+-----
+*2020-04-02*
+
+Features:
+- Major changes to config dialog:
+-- Set request method: GET and POST, with request body for POST
+-- Add request headers.
+-- Testing of feed config within dialog.
+-- Change polling interval.
+- Refresh button in menu
+- Feed error message displayed in menu.
+- Replaced RSS feature `with_url` with direct usage of `reqwest`
+- Better error-handling when polling feed.
+
+
 0.2.2
 -----
 *2020-03-14*
@@ -26,7 +43,6 @@ Notes:
 0.2.0
 -----
 *2019-10-12*
-
 
 Features:
 - Major restructuring.